Модуль 1 урок 2 задание 3

+++ module_1/lesson_2/src/solution_2.py
@@ -4,4 +4,4 @@ number_2 = int(input("Введите первое число - "))
 number_3 = int(input("Введите первое число - "))
 
 number_sum = int(number_1 + number_2 + number_3)
-print("Среднее арифметическое " + str(int(number_sum / 3)))
+print("Среднее арифметическое " + str(number_sum / 3))
